+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [6.0.4] - 2019-07-18
+- fix: Draw columns properly in readOnly mode.
 - fix: Add try around discussion restore selection.
 
 ## [6.0.3] - 2019-07-18